requirements.txt
farm-haystack==1.17.0
openpyxl==3.1.2
gradio==3.32.0
streamlit==1.22.0